instrumentation/opentelemetry-instrumentation-arangodb/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentelemetry-instrumentation-arangodb"
dynamic = ["version"]
description = "OpenTelemetry Arangodb instrumentation"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
  { name = "OpenTelemetry Authors", email = "cncf-opentelemetry-contributors@lists.cncf.io" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "opentelemetry-api ~= 1.27",
  "opentelemetry-instrumentation == 0.59b0.dev",
  "opentelemetry-semantic-conventions == 0.59b0.dev",
  "typing_extensions ~= 4.1",
]

[project.optional-dependencies]
instruments = ["python-arango >= 8.0, < 9.0"]

[project.entry-points.opentelemetry_instrumentor]
arangodb = "opentelemetry.instrumentation.arangodb:ArangoDBInstrumentor"

[project.urls]
Homepage = "https://github.com/open-telemetry/opentelemetry-python-contrib/tree/main/instrumentation/opentelemetry-instrumentation-arangodb"
Repository = "https://github.com/open-telemetry/opentelemetry-python-contrib"

[tool.hatch.version]
path = "src/opentelemetry/instrumentation/arangodb/version.py"

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/opentelemetry"]
